Potential implementation of the zeroConf framework
Changes:
- Created NsdHelper class which contains all the methods required for
perfoming network discovery, using NSD API. (Note, I still kept Frank's
classes for listeners, they look neat, might use those in the future)
- Created a method inside ServerCollectionInteractor called
discoverServer(), which should be called inside the UI in a AsyncTask
(should not be called on the main  thread).

How to proceed next: As discussed on Sunday during the group meeting at
Lars' ,  we decided to include this feature inside the IpDialog, as a
separate button. Also, when logging in, the user should first be
connected to the server, then login, therefore, this Dialog should be
popped up when someone uses the App for the first time.