Nitpicks + Fixed bugs causing the App to crash when the Context was null
Whe you close a window (Activity) before the AsyncTask was finished, it
attempted to print a message on the screen. To do so, it tried to get
the resources from the getContext(), which was null, causing the App to
crash.

Extracting the resources from the HestiaApplication itself will prevent
this from happening.